+++ Beam/config/database.php
@@ -131,8 +131,8 @@ return [
             'url' => env('REDIS_URL'),
             'host' => env('REDIS_HOST', '127.0.0.1'),
             'password' => env('REDIS_PASSWORD', null),
-            'port' => env('REDIS_PORT', 6379),
-            'database' => env('REDIS_DEFAULT_DATABASE', 0),
+            'port' => env('REDIS_PORT', '6379'),
+            'database' => env('REDIS_DEFAULT_DATABASE', '0'),
         ],
 
         'session' => [
@@ -146,19 +146,17 @@ return [
             'url' => env('REDIS_URL'),
             'host' => env('REDIS_HOST', '127.0.0.1'),
             'password' => env('REDIS_PASSWORD', null),
-            'port' => env('REDIS_PORT', 6379),
-            'database' => env('REDIS_CACHE_DATABASE', 2),
+            'port' => env('REDIS_PORT', '6379'),
+            'database' => env('REDIS_CACHE_DATABASE', '2'),
         ],
-
 
         'queue' => [
             'url' => env('REDIS_URL'),
             'host'     => env('REDIS_HOST', '127.0.0.1'),
             'password' => env('REDIS_PASSWORD', null),
-            'port' => env('REDIS_PORT', 6379),
-            'database' => env('REDIS_QUEUE_DATABASE', 3),
+            'port' => env('REDIS_PORT', '6379'),
+            'database' => env('REDIS_QUEUE_DATABASE', '3'),
         ],
-
 
     ],