+++ Beam/database/migrations/2021_05_06_142318_create_tag_tag_categories_table.php
@@ -1,0 +1,35 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class CreateTagTagCategoriesTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('tag_tag_category', function (Blueprint $table) {
+            $table->increments('id');
+            $table->integer('tag_id')->unsigned();
+            $table->integer('tag_category_id')->unsigned();
+
+            $table->foreign('tag_id')->references('id')->on('tags');
+            $table->foreign('tag_category_id')->references('id')->on('tag_categories');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('tag_tag_category');
+    }
+}