+++ Sso/config/auth.php
@@ -16,6 +16,7 @@ return [
     'defaults' => [
         'guard' => 'jwtx',
         'passwords' => null,
+
         'sso_provider' => env('DEFAULT_SSO_PROVIDER'),
     ],
 
@@ -45,6 +46,7 @@ return [
         'api' => [
             'driver' => 'token',
             'provider' => 'users',
+            'hash' => false,
         ],
 
         'jwt' => [
@@ -107,6 +109,7 @@ return [
             'provider' => 'users',
             'table' => 'password_resets',
             'expire' => 60,
+            'throttle' => 60,
         ],
     ],
 
@@ -123,10 +126,23 @@ return [
     |   - value: Label of SSO provider used on login page.
     |
     */
-
+    
     'sso_providers' => [
         'google' => 'Google',
         // define additional providers in services.php and enable them here
     ],
+
+    /*
+    |--------------------------------------------------------------------------
+    | Password Confirmation Timeout
+    |--------------------------------------------------------------------------
+    |
+    | Here you may define the amount of seconds before a password confirmation
+    | times out and the user is prompted to re-enter their password via the
+    | confirmation screen. By default, the timeout lasts for three hours.
+    |
+    */
+
+    'password_timeout' => 10800,
 
 ];