+++ Sso/config/database.php
@@ -120,6 +120,8 @@ return [
 
     'redis' => [
 
+        // Default Laravel 6+ is phpredis, but to avoid adding additional dependency (PHP extension)
+        // we still use predis
         'client' => env('REDIS_CLIENT', 'predis'),
 
         'options' => [
@@ -131,34 +133,33 @@ return [
             'url' => env('REDIS_URL'),
             'host' => env('REDIS_HOST', '127.0.0.1'),
             'password' => env('REDIS_PASSWORD', null),
-            'port' => env('REDIS_PORT', 6379),
-            'database' => env('REDIS_DEFAULT_DATABASE', 0),
+            'port' => env('REDIS_PORT', '6379'),
+            'database' => env('REDIS_DEFAULT_DATABASE', '0'),
         ],
 
         'session' => [
             'url' => env('REDIS_URL'),
             'host' => env('REDIS_HOST', '127.0.0.1'),
             'password' => env('REDIS_PASSWORD', null),
-            'port' => env('REDIS_PORT', 6379),
-            'database' => env('REDIS_SESSION_DATABASE', 1),
+            'port' => env('REDIS_PORT', '6379'),
+            'database' => env('REDIS_SESSION_DATABASE', '1'),
         ],
 
         'cache' => [
             'url' => env('REDIS_URL'),
             'host' => env('REDIS_HOST', '127.0.0.1'),
             'password' => env('REDIS_PASSWORD', null),
-            'port' => env('REDIS_PORT', 6379),
-            'database' => env('REDIS_CACHE_DATABASE', 2),
+            'port' => env('REDIS_PORT', '6379'),
+            'database' => env('REDIS_CACHE_DATABASE', '2'),
         ],
 
         'queue' => [
             'url' => env('REDIS_URL'),
             'host' => env('REDIS_HOST', '127.0.0.1'),
             'password' => env('REDIS_PASSWORD', null),
-            'port' => env('REDIS_PORT', 6379),
-            'database' => env('REDIS_QUEUE_DATABASE', 3),
+            'port' => env('REDIS_PORT', '6379'),
+            'database' => env('REDIS_QUEUE_DATABASE', '3'),
         ],
-
 
     ],